requirements.txt
transformers
seaborn
pycocotools>=2.0.7
Pillow>=10.0.0
matplotlib>=3.8.0
numpy>=1.24.0
tqdm>=4.66.0